+++ ManuvrOS/Drivers/Sensors/SensorManager.h
