+++ ManuvrOS/Drivers/Sensors/VEML6070/VEML6070.h
@@ -22,7 +22,7 @@ limitations under the License.
 #ifndef __VEML6070_DRIVER_H__
 #define __VEML6070_DRIVER_H__
 
-#include <Drivers/SensorWrapper/SensorWrapper.h>
+#include <Drivers/Sensors/SensorWrapper.h>
 #include <Platform/Peripherals/I2C/I2CAdapter.h>
 
 /* Hardware-defined registers */
